+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @remixapp/runtime
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies [3b5e5da]
+- Updated dependencies [20c7c35]
+- Updated dependencies [20c7c35]
+  - @remixapp/sdk@0.3.0
+
 ## 0.2.0
 
 ### Patch Changes